Move test support code to its own directory
