+++ server/test/api.edge.test.ts
@@ -3,8 +3,8 @@ import {
   installTestDatabaseHooks,
   useServerForTests,
   ndjsonParse,
-} from "./lib";
-import "./matchers";
+} from "./support/lib";
+import "./support/matchers";
 import { getApiKeys } from "../src/config";
 import app from "../src/app";
 import { createLocation, getLocationById, updateAvailability } from "../src/db";
