+++ server/test/db.test.ts
@@ -1,5 +1,5 @@
-import { installTestDatabaseHooks } from "./lib";
-import "./matchers";
+import { installTestDatabaseHooks } from "./support/lib";
+import "./support/matchers";
 import {
   createLocation,
   getAvailabilityForLocation,
